Bugfix: extra carriage return was being printed between the minus sign and the integer if printing a negative number
Feature: Added required PROGRAM keyword to begin a program and a period symbol after the final END to end it.  The compiler now compiles a (trivially limited) set of real Pascal programs.

Feature: Added support for the WRITE statement.

Feature: Added support for compiling files (previously, text was hardcoded in the call to the tokenizer in main())