Feature: Added support for if/then statement and the "equals" relational operator
Bugfix: Addressed situation where write() and writeln() could display integer values on screen, but they would not properly pipe to file either when stdout or stderr were redirected.  This fix involves including nsm64.asm and linking it in.

New Bug: If the pascal code includes a write(), that text will pipe to a file correctly.  However, succeeding write() and writeln() statements will not.  Once a writeln() is executed, future write/writeln will pipe to a file correctly, with write() subject to this bug again.